feat：乘客下单（任务调度）

+++ model/src/main/java/com/atguigu/daijia/model/entity/dispatch/XxlJobLog.java
@@ -27,6 +27,6 @@ public class XxlJobLog extends BaseEntity {
 
 	@Schema(description = "耗时(单位：毫秒)")
 	@TableField("times")
-	private Integer times;
+	private Long times;
 
 }
